requirements.txt
Django>=3,<4
django-extensions
djangorestframework
ldap3==2.9.1